Support BIG numbers in the database
The informer metric barfs when querying this without a cast, because of
scientific notation

Casting it to numeric means that our Go library can understand it

+++ pkg/db/store.go
@@ -313,7 +313,7 @@ func (s *EventStore) GetCFEventCount() (int64, error) {
 	row := s.db.QueryRowContext(
 		ctx,
 		fmt.Sprintf(
-			`SELECT reltuples FROM pg_class WHERE relname = '%s';`,
+			`SELECT reltuples::numeric FROM pg_class WHERE relname = '%s';`,
 			CFAuditEventsTable,
 		),
 	)
